+++ Changelog.md
@@ -1,6 +1,13 @@
 # Changelog
 
 ## [Unreleased]
+
+## [3.10.2] - 2026-06-04
+
+### Fixed
+- Fixed error occuring while registering did using update
+- Fixed issue of undefined getting passed as appId
+- Fixed issue of token is not getting re-generated in usage page
 ## [3.10.0] - 2026-06-02
 
 ### Added